Moved all routes to seperate files. Create helper_function file. Modified server.py to reflect changes.
