+++ backend/recipe_routes.py
@@ -1,0 +1,6 @@
+import status
+from flask import jsonify, request, Blueprint
+from firebase_admin import firestore
+from helper_functions import try_connect_to_db, user_validation
+
+recipe_bp = Blueprint("recipe", __name__)
